requirements.txt
absl-py==0.14.1
aioredis==1.3.1
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.9b0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
cycler==0.10.0
daphne==2.5.0
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
filelock==3.3.0
Flask==1.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
msgpack==0.6.2
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.3.3
pathlib==1.0.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pymongo==3.12.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
service-identity==21.1.0
six==1.15.0
sqlparse==0.2.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
tomli==1.2.1
tqdm==4.62.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
zope.interface==5.4.0
